refinement of the user update method and documentation

+++ internal/services/user.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"github.com/CatLecter/gin_template/internal/repositories"
 	"github.com/CatLecter/gin_template/internal/schemes"
+	"github.com/CatLecter/gin_template/utils"
 	"github.com/gofiber/fiber/v2"
 	"github.com/google/uuid"
 	log "github.com/sirupsen/logrus"
@@ -21,12 +22,12 @@ func (srv *UserService) GetUserByUUID(ctx *fiber.Ctx) error {
 	userUUID, err := uuid.Parse(ctx.Query("uuid"))
 	if err != nil {
 		log.Errorf("Error parsing user UUID: %v", err.Error())
-		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{"result": "cannot parse user UUID"})
+		return ctx.Status(fiber.StatusBadRequest).JSON(utils.NewHTTPError("cannot parse user UUID"))
 	}
 	userResp, err := srv.repo.GetUserByUUID(&userUUID)
 	if err != nil {
 		return ctx.Status(fiber.StatusNotFound).JSON(
-			fiber.Map{"result": fmt.Sprintf("user with UUID=%v not found", userUUID)},
+			utils.NewHTTPError(fmt.Sprintf("user with UUID=%v not found", userUUID)),
 		)
 	}
 	return ctx.Status(fiber.StatusOK).JSON(userResp)
@@ -36,21 +37,21 @@ func (srv *UserService) CreateUser(ctx *fiber.Ctx) error {
 	user := schemes.UserRequest{}
 	if err := ctx.BodyParser(&user); err != nil {
 		log.Warnf("Error parsing body: %v", err.Error())
-		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{"result": "cannot parse JSON"})
+		return ctx.Status(fiber.StatusBadRequest).JSON(utils.NewHTTPError("cannot parse JSON"))
 	}
 	isExists, err := srv.repo.CheckUserByPhone(&user.Phone)
 	if err != nil {
-		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{"result": "error checking the user's existence"})
+		return ctx.Status(fiber.StatusBadRequest).JSON(utils.NewHTTPError("error checking the user's existence"))
 	}
 	if *isExists {
 		return ctx.Status(fiber.StatusUnprocessableEntity).JSON(
-			fiber.Map{"result": fmt.Sprintf("user with phone number %v already exists", user.Phone)},
+			utils.NewHTTPError(fmt.Sprintf("user with phone number %v already exists", user.Phone)),
 		)
 	}
 	userResp, err := srv.repo.CreateUser(&user)
 	if err != nil {
 		log.Warnf("Error when creating a user: %v", err.Error())
-		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{"result": "user cannot be created"})
+		return ctx.Status(fiber.StatusBadRequest).JSON(utils.NewHTTPError("user cannot be created"))
 	}
 	return ctx.Status(fiber.StatusOK).JSON(userResp)
 }
@@ -59,27 +60,35 @@ func (srv *UserService) UpdateUserByUUID(ctx *fiber.Ctx) error {
 	userUUID, err := uuid.Parse(ctx.Query("uuid"))
 	if err != nil {
 		log.Warnf("Error parsing user UUID: %v", err.Error())
-		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{"result": "cannot parse user UUID"})
+		return ctx.Status(fiber.StatusBadRequest).JSON(utils.NewHTTPError("cannot parse user UUID"))
+	}
+	isExistsUser, err := srv.repo.CheckUserByUUID(&userUUID)
+	if err != nil {
+		return ctx.Status(fiber.StatusBadRequest).JSON(utils.NewHTTPError("error checking the user's existence"))
+	}
+	if !*isExistsUser {
+		return ctx.Status(fiber.StatusUnprocessableEntity).JSON(
+			utils.NewHTTPError(fmt.Sprintf("user with uuid %v not found", userUUID)),
+		)
 	}
 	user := schemes.UserRequest{}
 	if err := ctx.BodyParser(&user); err != nil {
 		log.Warnf("Error parsing body: %v", err.Error())
-		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{"result": "cannot parse JSON"})
+		return ctx.Status(fiber.StatusBadRequest).JSON(utils.NewHTTPError("cannot parse JSON"))
 	}
-
-	isExists, err := srv.repo.CheckUserByPhone(&user.Phone)
+	isExistsPhone, err := srv.repo.CheckUserByPhone(&user.Phone)
 	if err != nil {
-		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{"result": "error checking the user's existence"})
+		return ctx.Status(fiber.StatusBadRequest).JSON(utils.NewHTTPError("error checking the user's existence"))
 	}
-	if *isExists {
+	if *isExistsPhone {
 		return ctx.Status(fiber.StatusUnprocessableEntity).JSON(
-			fiber.Map{"result": fmt.Sprintf("user with phone number %v already exists", user.Phone)},
+			utils.NewHTTPError(fmt.Sprintf("user with phone number %v already exists", user.Phone)),
 		)
 	}
 	responseUser, err := srv.repo.UpdateUserByUUID(&userUUID, &user)
 	if err != nil {
 		log.Warnf("Error when updating a user: %v", err.Error())
-		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{"result": "user cannot be updated"})
+		return ctx.Status(fiber.StatusBadRequest).JSON(utils.NewHTTPError("user cannot be updated"))
 	}
 	return ctx.Status(fiber.StatusOK).JSON(responseUser)
 }
@@ -88,13 +97,22 @@ func (srv *UserService) DeleteUserByUUID(ctx *fiber.Ctx) error {
 	userUUID, err := uuid.Parse(ctx.Query("uuid"))
 	if err != nil {
 		log.Warnf("Error parsing user UUID: %v", err.Error())
-		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{"result": "cannot parse user UUID"})
+		return ctx.Status(fiber.StatusBadRequest).JSON(utils.NewHTTPError("cannot parse user UUID"))
+	}
+	isExistsUser, err := srv.repo.CheckUserByUUID(&userUUID)
+	if err != nil {
+		return ctx.Status(fiber.StatusBadRequest).JSON(utils.NewHTTPError("error checking the user's existence"))
+	}
+	if !*isExistsUser {
+		return ctx.Status(fiber.StatusNotFound).JSON(
+			utils.NewHTTPError(fmt.Sprintf("user with uuid %v not found", userUUID)),
+		)
 	}
 	err = srv.repo.DeleteUserByUUID(&userUUID)
 	if err != nil {
 		return ctx.Status(fiber.StatusNotFound).JSON(
-			fiber.Map{"result": fmt.Sprintf("user with an UUID=%v cannot be deleted", userUUID)},
+			utils.NewHTTPError(fmt.Sprintf("user with an UUID=%v cannot be deleted", userUUID)),
 		)
 	}
-	return ctx.Status(fiber.StatusOK).JSON(fiber.Map{"result": "success"})
+	return ctx.Status(fiber.StatusOK).JSON(schemes.HTTPResponse{Result: "success", Msg: "user deleted successfully"})
 }
